[skip ci] a lot changes...

+++ pre.sh
@@ -25,10 +25,19 @@ if [ ! -e m_gcc ] && [ ! -e m_binutils ]; then
 echo Unkown Error !; exit 255
 fi
 
+
+if [ -e x86_64-linux-gnu ]; then
+echo "Hmm? Huh!"
+else
+if [ x$NO_TC_DOWN = x ] && ( [ ! -e /opt/newcc ] || [ ! -e ~/x86_64-linux-gnu ] ); then
 curl -L "https://github.com/eebssk1/aio_tc_build/releases/download/NG-01/x86_64-linux-gnu-native.tgz" | tar -zxf -
 mv x86_64-linux-gnu /opt/newcc
+fi
+fi
 
 rm -rf m_*/build
 rm -rf out
 rm -rf *.tgz
 rm -rf time-*
+rm -rf tmp
+rm -rf x86_64-linux-gnu*
